requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
g2p-en==2.1.0
g2pM==0.1.2.5
huggingface-hub==0.30.2
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.4
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
praatio==6.2.0
propcache==0.3.1
pyarrow==19.0.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
TextGrid==1.6.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.19.0


#test